requirements.txt
fastapi==0.111.0
sqlalchemy==2.0.30
pydantic-settings==2.3.0
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3
pytest==8.2.2
pre-commit==3.7.1
pyjwt==2.8.0
passlib[bcrypt]==1.7.4